+++ pipenv/core.py
@@ -99,24 +99,6 @@ if PIPENV_COLORBLIND:
     crayons.disable()
 
 
-UNICODE_TO_ASCII_TRANSLATION_MAP = {
-    8230: u"...",
-    8211: u"-"
-}
-
-
-def fix_utf8(text):
-    if not isinstance(text, six.string_types):
-        return text
-    from ._compat import decode_output
-    try:
-        text = decode_output(text)
-    except UnicodeDecodeError:
-        if six.PY2:
-            text = unicode.translate(vistir.misc.to_text(text), UNICODE_TO_ASCII_TRANSLATION_MAP)
-    return text
-
-
 @contextlib.contextmanager
 def _spinner(text=None, nospin=None, spinner_name=None):
     if not text:
